+++ warehouse/forklift/forms.py
@@ -13,12 +13,103 @@
 import cgi
 import re
 
+import packaging.utils
 import wtforms
 import wtforms.validators
 
 
 from warehouse import forms
 from warehouse.forklift.metadata.form import ProjectName
+
+# Wheel platform checking
+
+# Note: defining new platform ABI compatibility tags that don't
+#       have a python.org binary release to anchor them is a
+#       complex task that needs more than just OS+architecture info.
+#       For Linux specifically, the platform ABI is defined by each
+#       individual distro version, so wheels built on one version may
+#       not even work on older versions of the same distro, let alone
+#       a completely different distro.
+#
+#       That means new entries should only be added given an
+#       accompanying ABI spec that explains how to build a
+#       compatible binary (see the manylinux specs as examples).
+
+# These platforms can be handled by a simple static list:
+_allowed_platforms = {
+    "any",
+    "win32",
+    "win_arm64",
+    "win_amd64",
+    "win_ia64",
+    "manylinux1_x86_64",
+    "manylinux1_i686",
+    "manylinux2010_x86_64",
+    "manylinux2010_i686",
+    "manylinux2014_x86_64",
+    "manylinux2014_i686",
+    "manylinux2014_aarch64",
+    "manylinux2014_armv7l",
+    "manylinux2014_ppc64",
+    "manylinux2014_ppc64le",
+    "manylinux2014_s390x",
+    "linux_armv6l",
+    "linux_armv7l",
+}
+# macosx is a little more complicated:
+_macosx_platform_re = re.compile(r"macosx_(?P<major>\d+)_(\d+)_(?P<arch>.*)")
+_macosx_arches = {
+    "ppc",
+    "ppc64",
+    "i386",
+    "x86_64",
+    "arm64",
+    "intel",
+    "fat",
+    "fat32",
+    "fat64",
+    "universal",
+    "universal2",
+}
+_macosx_major_versions = {
+    "10",
+    "11",
+    "12",
+    "13",
+    "14",
+}
+
+# manylinux pep600 and musllinux pep656 are a little more complicated:
+_linux_platform_re = re.compile(r"(?P<libc>(many|musl))linux_(\d+)_(\d+)_(?P<arch>.*)")
+_jointlinux_arches = {
+    "x86_64",
+    "i686",
+    "aarch64",
+    "armv7l",
+    "ppc64le",
+    "s390x",
+}
+_manylinux_arches = _jointlinux_arches | {"ppc64"}
+_musllinux_arches = _jointlinux_arches
+
+
+# Actual checking code;
+def _valid_platform_tag(platform_tag):
+    if platform_tag in _allowed_platforms:
+        return True
+    m = _macosx_platform_re.match(platform_tag)
+    if (
+        m
+        and m.group("major") in _macosx_major_versions
+        and m.group("arch") in _macosx_arches
+    ):
+        return True
+    m = _linux_platform_re.match(platform_tag)
+    if m and m.group("libc") == "musl":
+        return m.group("arch") in _musllinux_arches
+    if m and m.group("libc") == "many":
+        return m.group("arch") in _manylinux_arches
+    return False
 
 
 _dist_file_re = re.compile(r".+?(?P<extension>\.(tar\.gz|zip|whl))$", re.I)
@@ -43,8 +134,6 @@ class FileField(wtforms.FileField):
 
 
 def _validate_filename(form, field):
-    # Ensure the filename doesn't contain any characters that are too 🌶️spicy🥵
-
     # Our object storage does not tolerate some specific characters
     # ref: https://www.backblaze.com/b2/docs/files.html#file-names
     #
@@ -74,6 +163,24 @@ def _validate_filename(form, field):
         )
 
 
+def _validate_wheel_platform_tags(form, field):
+    # If this isn't a wheel, then we skip this validator
+    if not field.data.endswith(".whl"):
+        return
+
+    try:
+        _, _, _, tags = packaging.utils.parse_wheel_filename(field.data)
+    except packaging.utils.InvalidWheelFilename as exc:
+        raise wtforms.validators.ValidationError(str(exc))
+
+    for tag in tags:
+        if not _valid_platform_tag(tag.platform):
+            raise wtforms.validators.ValidationError(
+                f"Binary wheel {field.data!r} has an unsupported "
+                f"platform tag {tag.platform!r}"
+            )
+
+
 def _validate_filename_for_filetype(filename, filetype):
     if m := _dist_file_re.match(filename):
         extension = m.group("extension")
@@ -85,6 +192,13 @@ def _validate_filename_for_filetype(filename, filetype):
             )
 
 
+# NOTE: This form validation runs prior to ensuring that the current identity
+#       is authorized to upload for the given project, so it should not validate
+#       against anything other than what the user themselves have provided.
+#
+#       Any additional validations (such as duplicate filenames, etc) should
+#       occur elsewhere so that they can happen after we've authorized the request
+#       to upload for the given project.
 class UploadForm(forms.Form):
     # This field is duplicated out of the general metadata handling, to be part
     # of the upload form as well.
@@ -137,7 +251,11 @@ class UploadForm(forms.Form):
             # We purposely use DataRequired here, because we want to have this
             # work on coerced field data, not on the input data.
             wtforms.validators.DataRequired(),
+            # Ensure the filename doesn't contain any characters that are
+            # too 🌶️spicy🥵
             _validate_filename,
+            # Check that if it's a binary wheel, it's on a supported platform
+            _validate_wheel_platform_tags,
         ],
     )
 
@@ -163,6 +281,8 @@ class UploadForm(forms.Form):
 
         # We *must* have at least one digest to verify against.
         if (
+            # TODO: Don't consider md5_digest as good enough of satisifying the
+            #       message digest requirement.
             not self.md5_digest.data
             and not self.sha256_digest.data
             and not self.blake2_256_digest.data
